refactor: use structured output for `generateDocsSuggestion`
- Introduced `generateDocsSuggestion` function to analyze migration reviews and update documentation.
- Added `docsSuggestionSchema` for structured validation of the generated output.
- Refactored `processGenerateDocsSuggestion` to utilize the new function and handle responses more effectively.
- Created an index file for easier exports from the `generateDocsSuggestion` module.
