+++ frontend/packages/jobs/src/functions/processGenerateDocsSuggestion.ts
@@ -1,77 +1,7 @@
-import { PromptTemplate } from '@langchain/core/prompts'
-import { ChatOpenAI } from '@langchain/openai'
 import { prisma } from '@liam-hq/db'
 import { getFileContent } from '@liam-hq/github'
+import { generateDocsSuggestion } from '../prompts/generateDocsSuggestion/generateDocsSuggestion'
 import { langfuseLangchainHandler } from './langfuseLangchainHandler'
-
-export const MIGRATION_DOCS_REVIEW_TEMPLATE = `
-You are Liam, an expert in schema design and migration strategy for this project.
-
-Your task is to analyze migration reviews and update the internal documentation files in docs/ to maintain accurate, structured, and reusable knowledge.
-
-🎯 Goal:
-Extract project-specific conventions, constraints, and patterns from the migration review that inform future schema design and migration practices.
-
-## 📁 Documentation Structure
-
-The following files need to be maintained:
-
-schemaPatterns.md:
-- Reusable patterns and rules for database schema design
-- Structural modeling patterns, naming conventions, preferred types
-- Canonical design choices specific to this project
-
-schemaContext.md:
-- Project-specific constraints that shape schema design
-- Technical assumptions, ORM limitations, domain modeling needs
-- Only schema-wide policies (no specific fields/models)
-
-migrationPatterns.md:
-- Safe and consistent migration strategies
-- Sequencing rules, rollout patterns, reversibility requirements
-- Implementation standards for this project
-
-migrationOpsContext.md:
-- Operational constraints on executing migrations
-- Timing, tooling, deployment risks, safety strategies
-
-.liamrules:
-- Informal but recurring knowledge
-- Field/model specific patterns
-- One-time decisions that may inform future work
-
----
-
-Migration Review:
-{reviewResult}
-
-Current Documentation:
-{docsArray}
-
----
-
-Please analyze the migration review and:
-1. Identify any new project-specific patterns or constraints
-2. Update relevant documentation files with new knowledge
-3. Add field/model specific insights to .liamrules
-4. Return the complete updated content for any modified files
-
-Return your analysis as a JSON object with the following keys:
-// schemaPatterns
-// schemaContext
-// migrationPatterns
-// migrationOpsContext
-// liamrules
-
-Each key should contain the full updated content for the respective file, or be left unchanged.
-
-Remember:
-- Only include project-specific insights
-- Be precise and intentional
-- Focus on reusable knowledge
-- Maintain accuracy and clarity
-- Return valid JSON format
-`
 
 const DOC_FILES = [
   'schemaPatterns.md',
@@ -123,38 +53,22 @@ export async function processGenerateDocsSuggestion(payload: {
     })
 
     const docsArray = await Promise.all(docsPromises)
+    const docsArrayString =
+      docsArray.length > 0
+        ? JSON.stringify(docsArray)
+        : 'No existing docs found'
 
-    const prompt = PromptTemplate.fromTemplate(MIGRATION_DOCS_REVIEW_TEMPLATE)
-
-    const model = new ChatOpenAI({
-      temperature: 0.7,
-      model: 'gpt-4o-mini',
-    })
-
-    const chain = prompt.pipe(model)
-    const response = await chain.invoke(
-      {
-        reviewResult: payload.reviewComment,
-        docsArray:
-          docsArray.length > 0
-            ? JSON.stringify(docsArray)
-            : 'No existing docs found',
-      },
-      {
-        callbacks: [langfuseLangchainHandler],
-      },
+    const callbacks = [langfuseLangchainHandler]
+    const result = await generateDocsSuggestion(
+      payload.reviewComment,
+      docsArrayString,
+      callbacks,
     )
 
-    // Clean up the response and parse JSON
-    try {
-      const rawContent = response.content.toString()
-      // Remove markdown code block markers if they exist
-      const jsonStr = rawContent.replace(/^```(?:json)?\s*|\s*```$/g, '')
-      return JSON.parse(jsonStr)
-    } catch (error) {
-      console.error('Error parsing LLM response:', error)
-      throw new Error('Invalid response format from LLM')
-    }
+    // Filter out undefined values and return
+    return Object.fromEntries(
+      Object.entries(result).filter(([_, value]) => value !== undefined),
+    ) as Record<string, string>
   } catch (error) {
     console.error('Error generating docs suggestions:', error)
     throw error
