+++ frontend/packages/jobs/src/prompts/generateDocsSuggestion/docsSuggestionSchema.ts
@@ -1,0 +1,9 @@
+import { object, string } from 'valibot'
+
+export const docsSuggestionSchema = object({
+  schemaPatterns: string(),
+  schemaContext: string(),
+  migrationPatterns: string(),
+  migrationOpsContext: string(),
+  liamrules: string(),
+})
