+++ frontend/packages/jobs/src/prompts/generateDocsSuggestion/index.ts
@@ -1,0 +1,2 @@
+export * from './docsSuggestionSchema'
+export * from './generateDocsSuggestion'
